rename libquix to libnitrate
