+++ libnitrate/fuzzer/libnitrate-basic-lexer-fuzz.cc
@@ -1,4 +1,4 @@
-#include <quix/code.h>
+#include <nitrate/code.h>
 
 extern "C" int LLVMFuzzerTestOneInput(const uint8_t *Data, size_t Size) {
   if (Size == 0) {
